add resources

+++ pkg/config/defaults.go
@@ -3,9 +3,9 @@ package config
 import "time"
 
 const (
-	DefaultIntegerRangeMin  uint32 = 0
-	DefaultIntegerRangeMax  uint32 = 10
-	DefaultIntegerRangeStep uint32 = 1
+	DefaultIntegerRangeMin  int64  = 0
+	DefaultIntegerRangeMax  int64  = 10
+	DefaultIntegerRangeStep int64  = 1
 	DefaultIntegerRangePad  uint32 = 0
 
 	DefaultRandomStringSize    uint32 = 10
